requirements.txt
streamlit==1.32.0
requests==2.31.0
python-docx==1.0.1
PyPDF2==3.0.1
appdirs==1.4.4
